+++ src/editor-model/commands.ts
@@ -334,9 +334,9 @@ export function move(
 
   if (pos < 0 || pos > model.lastOffset) {
     // We're going out of bounds
-    let result = true; // True => perform default handling
+    let success = true; // True => perform default handling
     if (!model.silenceNotifications) {
-      result =
+      success =
         model.mathfield.host?.dispatchEvent(
           new CustomEvent('move-out', {
             detail: { direction },
@@ -346,8 +346,8 @@ export function move(
           })
         ) ?? true;
     }
-    if (result) model.announce('plonk');
-    return result;
+    if (success) model.announce('plonk');
+    return success;
   }
 
   model.setPositionHandlingPlaceholder(pos);
@@ -473,9 +473,9 @@ function moveUpward(model: _Model, options?: { extend: boolean }): boolean {
 
   // Callback when there is nowhere to move
   const handleDeadEnd = () => {
-    let result = true; // True => perform default handling
+    let success = true; // True => perform default handling
     if (!model.silenceNotifications) {
-      result =
+      success =
         model.mathfield.host?.dispatchEvent(
           new CustomEvent('move-out', {
             detail: { direction: 'upward' },
@@ -485,8 +485,8 @@ function moveUpward(model: _Model, options?: { extend: boolean }): boolean {
           })
         ) ?? true;
     }
-    model.announce(result ? 'plonk' : 'line');
-    return result;
+    model.announce(success ? 'line' : 'plonk');
+    return success;
   };
 
   // Find a target branch
@@ -545,9 +545,9 @@ function moveDownward(model: _Model, options?: { extend: boolean }): boolean {
   if (!extend) model.collapseSelection('forward');
   // Callback when there is nowhere to move
   const handleDeadEnd = () => {
-    let result = true; // True => perform default handling
+    let success = true; // True => perform default handling
     if (!model.silenceNotifications) {
-      result =
+      success =
         model.mathfield.host?.dispatchEvent(
           new CustomEvent('move-out', {
             detail: { direction: 'downward' },
@@ -557,8 +557,8 @@ function moveDownward(model: _Model, options?: { extend: boolean }): boolean {
           })
         ) ?? true;
     }
-    model.announce(result ? 'plonk' : 'line');
-    return result;
+    model.announce(success ? 'line' : 'plonk');
+    return success;
   };
 
   // Find a target branch